🔨 refactor: id 타입 long으로 변경

+++ src/main/java/com/x1/frans/supplier/command/application/service/SupplierCommandService.java
@@ -4,6 +4,6 @@ import com.x1.frans.supplier.command.vo.SupplierUpdateRequestVO;
 
 public interface SupplierCommandService {
 
-    void updateSupplier(int userId, int supplierId, SupplierUpdateRequestVO supplierUpdateRequestVO);
+    void updateSupplier(long userId, long supplierId, SupplierUpdateRequestVO supplierUpdateRequestVO);
 
 }
